Edit Image Optional navigation Add note optional changes for image
